refactor<search>搜索功能的分页

+++ core/data/src/main/java/com/example/data/model/search/detail/SearchDetail.kt
@@ -1,7 +1,5 @@
 package com.example.data.model.search.detail
 
-import java.util.function.LongFunction
-
 data class SearchDetail(
     val code: Int,
     val result: Result,
@@ -174,7 +172,7 @@ data class PlayLists(
     val specialType: Long,
     val subscribed: Boolean,
     val track: Track,
-    val trackCount: Long,
+    val trackCount: Int,
     val userId: Long
 )
